Factor out JHasRunCallbacks
